GitCmdStreamReader: fixed terrible bug which only kicked in if the stream was actually empty. This is a rare case that can happen during stream testing. Theoretically there shouldn't be any empty streams of course, but practically they do exist sometimes ;); fixed stream.seek implementation, which previously used seek on standard output Improved GitCmd error handling

+++ test/testlib/helper.py
@@ -109,11 +109,14 @@ def with_rw_repo(working_tree_ref):
 					return func(self, rw_repo)
 				except:
 					print >> sys.stderr, "Keeping repo after failure: %s" % repo_dir
+					repo_dir = None
 					raise
 			finally:
 				os.chdir(prev_cwd)
 				rw_repo.git.clear_cache()
-				shutil.rmtree(repo_dir, onerror=_rmtree_onerror)
+				if repo_dir is not None:
+					shutil.rmtree(repo_dir, onerror=_rmtree_onerror)
+				# END rm test repo if possible
 			# END cleanup
 		# END rw repo creator
 		repo_creator.__name__ = func.__name__
